requirements.txt
google-generativeai>=0.7.0
openai>=1.0.0
rapidfuzz>=3.6.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0
colorama>=0.4.6
tabulate>=0.9.0
